# doc/api.yaml
openapi: 3.0.3
servers:
  - url: "http://localhost:3000"
info:
  title: WASAText
  description: |-
   Connect with your friends effortlessly using WASAText!
   Send and receive messages, whether one-on-one or in groups,   all from
   the convenience of your PC. Enjoy seamless conversations with text or
   GIFs and easily stay in touch through your private chats or group
   discussions.
  version: 1.0.0
security:
  - BearerAuth: []
#18---------------------------------------------------------------------------------------------------
paths:
...
  /users:
    get:
      operationId: searchUsers
      parameters:
        - in: query
          name: searchkey
          required: false
          schema:
            type: string
      responses:
        '200':
          description: user list found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        '204':
          description: 'No users found with this key'
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /profile/username:
    put:
      tags: ["username setting"]
      operationId: setMyUserName
      description: |-
        Modifies current username to another unique alphanum string
        chosen by the user
      requestBody:
        required: true
        content:
            application/json:
              schema:
                type: object
                required:
                  - targetusername
                properties:
                  targetusername:
                    type: string
                    example: Maria03
                    minLength: 3
                    maxLength: 16
                    pattern: '^[a-zA-Z0-9]+$'
      responses:
        '204':
          description: 'username change successful'
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }
        "409":
          description: username already in use
  /profile/photo:
    put:
      tags: ["pfp setting"]
      operationId: setMyPhoto
      description: |-
        Modifies current profile photo to another. When user deletes
        their own photo/chooses not to have one, use a black or white
        picture
      requestBody:
        content:
            image/jpeg:
              schema:
                type: string
                format: binary
        required: true
      responses:
        '204':
          description: 'photo change successful'
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /chats/groups: 
    post: 
      operationId: makeGroup
      description: |-
        makes a new group, returns a chat.
        requesting user added to members by api handler
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              type: object
              required:
                - members
                - groupname
              properties:
                members:
                  type: array
                  items:
                    type: string
                groupname:
                  type: string
      responses:
        '201':
          description: made a new group
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Chat"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '500': {$ref: "#/components/responses/InternalServerError"}
...
  /chats/groups/{chatID}/photo:
    put: 
      operationId: setGroupPhoto
      requestBody:
        required: true
        content:
          image/jpeg:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: groupchat picture updated
        '400': {$ref: "#/components/responses/BadRequest"}
        '403':
          description: only admins can change the groupchat profile picture
        '404':
          description: groupchat does not exist
        '500': {$ref: "#/components/responses/InternalServerError"}
    parameters:
      - $ref: "#/components/parameters/chatID"
  /chats/groups/{chatID}/groupname:
    put: 
      operationId: setGroupName
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
              minLength: 1
              maxLength: 50
      responses:
        '200':
          description: groupchat name updated
        '400': {$ref: "#/components/responses/BadRequest"}
        '403':
          description: only admins can change the groupchat name
        '500': {$ref: "#/components/responses/InternalServerError"}
    parameters:
      - $ref: "#/components/parameters/chatID"
  /chats/{chatID}/messages:
    get:
      operationId: getConversation
      description: returns all msgs in reverse chronological order
      responses:
        '200':
          description: convo retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
        '400': {$ref: "#/components/responses/BadRequest"}
        '500': {$ref: "#/components/responses/InternalServerError"}

    post:
      operationId: sendMessage
      description: sends a message containing text and/or an image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              oneOf:
                - $ref: "#/components/schemas/SendMessageReq"
                - $ref: "#/components/schemas/SendPhotoMessageReq"
      responses:
        '201':
          description: message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

        '400': {$ref: "#/components/responses/BadRequest"}
        '500': {$ref: "#/components/responses/InternalServerError"}
    parameters: 
    - $ref: "#/components/parameters/chatID"
...
  /chats/{chatID}/messages/{messageID}/reactions/{reactionID}:
    delete:
      operationId: uncommentMessage
      responses:
        '204':
          description: reaction removed
        '400': {$ref: "#/components/responses/BadRequest"}
        '403':
          description: user is not the owner of this reaction
        '500': {$ref: "#/components/responses/InternalServerError"}

    parameters:
      - $ref: "#/components/parameters/chatID"
      - $ref: "#/components/parameters/messageID"
      - $ref: "#/components/parameters/reactionID"
components:
  schemas:
    SendMessageReq:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 20000
        replyTo:
          type: string
    SendPhotoMessageReq:
      type: object
      required:
        - photo
      properties:
        photo:
          type: string
          format: binary
        replyTo:
          type: string
    Chat:
      type: object
      required:
        - chatID
        - groupORchat
        - chatName
        - members
        - creationDate
      properties:
        chatID:
          type: string
        groupORchat:
          type: string
          enum:
            - private
            - group
        chatName:
          type: string
          description: username if private, groupname if group
        members:
          type: array
          items:
            type: string
        creationDate:
          type: string
          format: date-time
        photo:
          description: the profile picture of the user you are chatting with, or an image chosen for the group
          type: string
          format: uri
        mostrecentmsg:
          $ref: "#/components/schemas/MessagePreview"
    User:
      title: user
      description: |-
        a generic user: can refer to oneself, or to other persons
        perusing the app from any device.
      type: object
      required:
        - username
        - userID
      properties:
        photo:
          description: the profile picture of the user you are chatting with
          type: string
          format: uri
        username:
          description: unique username, alphanumeric
          type: string
          example: Maria02
          minLength: 3
          maxLength: 16
          readOnly: true
        userID:
          type: string
          example: ggfvg255
    Message:
      title: message
      description: a single message from a user at a specific date-time
      type: object
      required:
        - messageID
        - senderID
        - status
        - timestamp
      oneOf: 
        - required:
          - content
        - required:
          - photo
      properties:
        messageID:
          type: string
        senderID:
          description: userID of the msg sender
          type: string
        content:
          description: |-
            the word contents of the message, can be none if
            sending a picture
          type: string
        photo:
          type: string
          format: uri
        status:
          type: string
          enum:
            - sent
            - delivered
            - read
        timestamp:
          type: string
          format: date-time
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/Reaction"
        replyTo:
          type: string
          description: |-
            must contain the msg ID you're replying to. needed
            for backend to load msg preview!
        fwdFrom:
          type: string
          description: ID of the og msg that was fwd
    MessagePreview:
      title: message
      description: a single message from a user at a specific date-time
      type: object
      required:
        - messageID
        - timestamp
        - contentType
        #can't use content and photo
      properties:
        senderID:
          type: string
        content:
          description: text of msg, absent when msg is img-only
          type: string
        contentType:
          type: string
          enum:
            - text
            - photo
            - text photo
        timestamp:
          type: string
          format: date-time
        messageID:
          type: string
    Reaction:
      description: |-
        an emoticon added by a user (including oneself) at the foot of a message
        feature: one reaction per user per message
      type: object
      required:
        - emoji
        - userID
        - reactionID
      properties:
        userID:
          type: string
        emoji:
          type: string
        reactionID:
          type: string
...
  responses:
...
    BadRequest:
      description: request not compliant with documentation
    InternalServerError:
      description: server had an internal error. More info in server logs
  securitySchemes:
      BearerAuth:
        type: http
        scheme: bearer
  parameters:
    chatID:
      name: chatID
      in: path
      required: true
      schema:
        type: string
        format: uuid
    messageID:
      name: messageID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    reactionID:
      name: reactionID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...